refactor(swapStrategies) move to own file;
