+++ src/swap-strategies.js
@@ -1,0 +1,28 @@
+export const SwapStrategies = {
+  // animate the next view in before removing the current view;
+  before(viewSlot, previousView, callback) {
+    let promise = Promise.resolve(callback());
+
+    if (previousView !== undefined) {
+      return promise.then(() => viewSlot.remove(previousView, true));
+    }
+
+    return promise;
+  },
+
+  // animate the next view at the same time the current view is removed
+  with(viewSlot, previousView, callback) {
+    let promise = Promise.resolve(callback());
+
+    if (previousView !== undefined) {
+      return Promise.all([viewSlot.remove(previousView, true), promise]);
+    }
+
+    return promise;
+  },
+
+  // animate the next view in after the current view has been removed
+  after(viewSlot, previousView, callback) {
+    return Promise.resolve(viewSlot.removeAll(true)).then(callback);
+  }
+};
